Fixing issue #3
When a getter or putter future was cancelled from the outside but not
yet removed from the _getters / _putters deque, and then calling close(),
an InvalidStateError was thrown when blindly calling set_exception before
checking if the future was cancelled.

I've previously used future.done() check here, but I don't want that,
because I _want_ this code to fail if a _value_ has been previously set
on these futures. However, as @stijn-devriendt corrected pointed out,
these futures can be cancelled from the outside, which should be a valid
operation.

I've implemented a check to see if a putter/getter is already cancelled
before setting the ChannelClosed exception on close().

I've also considered what happens if we have a lot of putters/getters
in the deque that has been cancelled. I could remove them by hooking up
a done callback to the futures, but I don't think this necesary because
if a putter/getter is already marked done, it will simply be popped and skipped
when it is it's turn in the deque, so cleaning up will happen automatically
once the cancelled future reaches the head of the deque.

For almost all use-cases I can think of, that is both good enough, and probably
also usually better as to not have to search a large deque every time a future
is cancelled to make sure we remove it.