Fixes #13
`close` will mark the channel as closed, but only cancel getters that _cannot_
be satisfied by the remaining items in the channel.

The problem is that nobody ever told the remaining getters that they could proceed,
leading to waiting forever.

This fix will wake up the remaining getters, such that when `close` completes,
there should be no more pending getters or putters.

New getters can ofc. arrive later as normal.